requirements.txt
numpy
scipy
pandas
mne
torch
scikit-learn
matplotlib
tqdm
tensorflow-cpu  # or latest supported version
streamlit
